+++ src/nymphes_midi/NymphesMIDI.py
@@ -936,6 +936,44 @@ class NymphesMIDI:
             # Send the preset to Nymphes and connected MIDI Output ports
             self._preset_snapshot_needed = True
 
+    def load_syx_file(self, filepath):
+        """
+        Load the .syx file at filepath
+        :param filepath: A Path or string. The path to the preset file.
+        :return:
+        """
+        if self.nymphes_connected:
+            try:
+                # Load the file into a list of MIDI messages
+                messages = mido.read_syx_file(filepath)
+
+                # Get the first message
+                msg = messages[0]
+
+                # Get its data
+                msg_data = msg.data
+
+                # Load the preset file as the current preset
+                self._curr_preset_object = NymphesPreset(sysex_data=msg_data)
+
+                # Reset the unsaved changes flag
+                self._unsaved_changes = False
+
+                # Notify Client
+                self.add_notification(
+                    PresetEvents.loaded_file.value,
+                    str(filepath)
+                )
+
+                # Send all parameters to the client
+                self.send_current_preset_notifications()
+
+                # Send the preset to Nymphes and connected MIDI Output ports
+                self._preset_snapshot_needed = True
+
+            except Exception as e:
+                Logger.warning(f'Failed to load .syx file as a Nymphes preset: {filepath}, {e}')
+
     def load_init_file(self):
         """
         Load init.txt